Rename the common lib
