+++ web/src/common/layers/DynamicMovementArrows.svelte
