+++ web/src/common/layers/RenderIntersectionMarkings.svelte
