+++ web/src/common/layers/RenderIntersectionPolygons.svelte
