+++ web/src/common/osm_input/BuiltInSelector.svelte
